p2p/Rotation: Fix a potential error causing more nodes to leave than wanted
`maxRemove` is initialized to `maxRotatedPerCycle`, and if it's set to 1 or
more, the old code will override it with `scaleDownRemove`, which i'm
guessing can be more than `maxRotatedPerCycle` (i.e. 1).

if we wanted the amount of rotated nodes to be strictly equal to or less
than `maxRotatedPerCycle`, then `maxRotatedPerCycle` should always be
the cap (upper bound).