Add `SyncV2` module
This is a p2p module that contains all of the functionality for the new
Syncing Protocol v2.

The module index exposes a top-level `syncV2` function that handles all
of the synchronization, setting values in NodeList, Archivers, and
CycleChain as appropriate.

The `queries` submodule contains logic pertaining to anything that
requires an external node to be queried. This includes robust queries.

The `routes` submodule defines any routes that need to be exposed and
available for other nodes to query.

The new SyncV2 module as a whole uses the `neverthrow` library to ensure
that all errors are handled when they happen. You may notice the lack of
`async` and `await` keywords in this module in favor of the
`ResultAsync` type.

The `Result` types allow us to take an approach more akin to functional
programming when it comes to writing the logic of our code. We make
frequent use of the `andThen` method on `ResultAsync` (or sometimes
`asyncAndThen` on `Result`). It works a lot like how you would expect
any other boolean operation to work: if the `Result` is `Ok`, execution
continues and the passed function is called (which, in the case of
`andThen`, returns another `Result` type that is handled next in the
chain). But if the `Result` is `Err`, the chain short-circuits and only
the `Err` is returned.

The readability and practicality of this code may be subjective, but I
personally like the avoidance of any uncaught errors and littering
try-catch blocks everywhere. I believe it is more understandable to read
e.g. "syncValidValidatorList andThen syncValidArchiverList" and so on.