GOLD-172 Migrating PoQo endpoints to Binary (#172)
* feat: binary migration of popqo endpints

* PoQo Send Receipt Binary

Binary Migration of PoQo Send Receipt : Handler, Serializer and Deserializer Additions

* feat: Add poqoSendReceiptBinary configuration option

The `poqoSendReceiptBinary` configuration option is added to enable binary migration of the PoQo Send Receipt feature. This commit includes the necessary additions to the handler, serializer, and deserializer for the binary migration.

* refactor: Update internal route for sending receipt in TransactionConsensus.ts

* formatting endpoint position

* feat: Add sending final receipt gossip to binary handler in TransactionConsensus.ts

* refactor: Improve error handling in poqoSendVoteBinary handler

* feat: Fix error handling and improve binary migration in poqo  data and receipt

* refactor: Update poqoSendReceiptReq deserialization logic

Refactor the deserialization logic in the `deserializePoqoSendReceiptReq` function in the `PoqoSendReceiptReq.ts` file. Simplify the code by removing the unnecessary `else` block and handling the `confirmOrChallenge` property more efficiently.

* fix: redundant binary endpoint call

* fix: redundant serde functions and enable bin serializtion

---------

Co-authored-by: Kaung Myat Thu <2962000kaungmyatthu@gmail.com>
Co-authored-by: Atharva <atharvakn@gmail.com>